Day 1 - also updated template to utilise a shared main function.
I iterate through the input once tallying up calorie counts for each elf, and adding the value to an array when a blank line is encountered, resetting the tally back to 0.
I then use built in sorting/max/sum methods in order to return my result \o/

There's more efficient ways of doing this such as only remembering the highest 1 or 3 values and performing that check at each blank, but this method would be more extensible, allowing for other operations to be easily performed should the task had requested them!